backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
scikit-learn
xgboost
prophet
tensorflow
keras
nltk
vaderSentiment
sqlalchemy
pymysql
python-jose
passlib
bcrypt
python-dotenv
python-multipart
pydantic
pydantic-settings
httpx
celery
redis
matplotlib
plotly